add afl-clang-fast debug dump
It turns out that afl-clang-fast clobbers clang args in ways that do not
mesh with how one would normally pass args to clang. To debug this mess,
add a dump method that allows us to see the before/after view of the
args, to try and pinpoint where things go wrong.